requirements.txt
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.3
tenacity>=8.2.3
python-dotenv>=1.0.1
